Move license match functions to licensing
